Calculate the send window dynamically.
Rather than calculate the send_window and then storing it, we can
derive it by a calculation on the retransmit queue. This is more safe,
albeit it is a bit slower than the solution we had before.

The problem however, is that the send_window_packets count was not
updated correctly and that it is really an optimization in the code.